fix some bugs

+++ predict.py
@@ -71,7 +71,7 @@ def predict_postprocess(batch_idx,posprocess,batch_pred_masks,shape=(350,525),mi
 
         if posprocess:
             pred_masks = batch_pred_masks[j,]
-            pred_masks = cv2.resize(pred_masks, dsize=(w, h), interpolation=cv2.INTER_LINEAR)
+            pred_masks = cv2.resize(np.float32(pred_masks), dsize=(w, h), interpolation=cv2.INTER_LINEAR)
             arrt = np.array([])
             for t in range(4):
                 a, num_predict = post_process(sigmoid(pred_masks[:, :, t]),thresholds[t], minsizes[t], shape)
